Add new environment type
Add EnvironmentBuffering safe and unsafe (incomplete, still needs some
implementation)

+++ solver/src/main/java/choco/kernel/memory/copy/EnvironmentCopying.java
@@ -366,7 +366,7 @@ public class EnvironmentCopying extends AbstractEnvironment {
      * {@inheritDoc}
      */
     @Override
-    public IStateLong makeLong(int init) {
+    public IStateLong makeLong(long init) {
         return new RcLong(this, init);
     }
